requirements.txt
pandas
numpy
matplotlib
seaborn
gradio
kaggle
kagglehub
Pillow
# PyTorch ekleniyor
torch
torchvision